v2.5
Successfully connected assets and account page. Data submitted on assets page is saved to sql db and then outputted in asset breakdown on account page. Next steps are to do the same for investments and debts page, and to be able to edit and delete each record. Further steps are to use the breakdown information to display an overall net worth.